# ARROW-10703: [Rust] [DataFusion] Make join not collect on every right part