This is a draft proposal to address a limitation on which the join currently computes the left side (build) on every part of the right (probe).

@alamb and @andygrove , I am not sure if this is right from the execution's point of view. the idea here is to leverage the notion that as long as it has all the information it requires, a node can always return all its streams simply by iterating on them. We can scrap this if it does not feel right.

Note that this makes `Merge` to not `Plan::execute` inside of a `spawn`. I am not sure if this has non trivial consequences.