@Dandandan 2020-12-13
@jorgecarleitao 
I checked out the branch locally and ran it, the implementation unfortunately causes a ~40x slowdown on a query that includes a join (I double checked that I was running in release mode):

Query 12 iteration 0 took 51340 ms
Query 12 iteration 1 took 53788 ms
Query 12 iteration 2 took 53361 ms