@jorgecarleitao 2020-12-14
Well, clearly this does not work :)

I do not understand why, but this is not ready to merge.

FWIW, IMO this issue (not the PR) is a priority for the 3.0.0 release.